Spam-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
werkzeug==2.3.7
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.4
python-dotenv==1.0.0
numpy==1.24.4
requests==2.31.0
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
bcrypt==4.1.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
